% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeWrite.R
\name{initializeWrite}
\alias{initializeWrite}
\title{Start the writing process}
\usage{
initializeWrite(dir = tempfile())
}
\arguments{
\item{dir}{String containing the path to a staging directory.
Any existing file at that location is deleted and \code{dir} is created.}
}
\value{
String containing a path to a configured staging directory.
}
\description{
Set up the staging directory and state file for creating a \pkg{kana} file.
}
\examples{
dir <- initializeWrite()
list.files(dir)

}
\author{
Aaron Lun
}
