% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeQualityControl.R
\name{writeQualityControl}
\alias{writeQualityControl}
\title{Write contents for quality control}
\usage{
writeQualityControl(
  dir,
  metrics,
  thresholds,
  discards,
  use_mito_default = TRUE,
  mito_prefix = "mt-",
  nmads = 3
)
}
\arguments{
\item{dir}{String containing the working directory for preparing the \pkg{kana} output.}

\item{metrics}{List containing \code{sums}, \code{detected} and \code{proportions},
which are numeric, integer and numeric vectors, respectively, of length equal to the number of cells.}

\item{thresholds}{List containing \code{sums}, \code{detected} and \code{proportions}, 
all of which are numeric vectors of length equal to the number of samples.}

\item{discards}{Logical vector of length equal to the number of cells.}

\item{use_mito_default, mito_prefix, nmads}{Parameters for the \pkg{kana} analysis.}
}
\value{
The QC contents are written to file.
A \code{NULL} is invisibly returned.
}
\description{
Write quality control-related contents to the HDF5 state file.
}
\details{
See \url{https://ltla.github.io/kanaval/quality__control_8hpp.html} for details.
}
\examples{
dir <- initializeWrite()
writeQualityControl(dir, 
    metrics = list(
        sums = runif(100), 
        detected = sample(1000, 100), 
        proportions = runif(100)
    ),
    thresholds = list(
        sums = runif(2), 
        detected = sample(1000, 2), 
        proportions = runif(2)
    ),
    discards = rbinom(100, 1, 0.5) == 0
)

}
\author{
Aaron Lun
}
